requirements.txt
discord
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto==2.49.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
dnspython==1.16.0
grpcio==1.28.1
httplib2==0.17.1
idna==2.9
git+https://github.com/Rapptz/discord.py
msgpack==1.0.0
multidict==4.7.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
